rename user find by id use case and controller, add new controller for reading user by id as client
